# Optimize count(distinct <complex type>)