Aggregations over distinct inputs use `SetAccumulator<ComplexType>`, which in
turn uses AddressableNonNullValueList to store unique complex type values in a
single non-contiguous allocation within HashStringAllocator.

Storing thousands or millions of values requires allocation with hundrends or
thousands of contiguous pieces. Calling HashStringAllocator::prepareRead on
such allocation ends up populating an `std::vector<ByteRange>` with thousands
of entries. Calling this repeatedly for each value as part of
SetAccumulator::extractValues() becomes very slow.

SetAccumulator::extractValues calls AddressableNonNullValueList::read for each
unique value (there can be millions of these):

    for (const auto& position : base.uniqueValues) {
      AddressableNonNullValueList::read(
          position.first.position, values, offset + position.second);
    }

AddressableNonNullValueList::read calls HashStringAllocator::prepareRead, which
collects thousands of byte ranges into a vector passed to ByteInputStream
constructor:

    auto stream = HashStringAllocator::prepareRead(header);

As a result, queries like count(distrinct <complex type>) are very slow.

A fix is to modify HashStringAllocator::prepareRead to accept an optional limit
on how many bytes to prepare for read, then use this in
AddressableNonNullValueList::read to prepare only as many bytes as needed to
extract a single value.

In addition, do not store hashes of the values in HSA to avoid calling
HashStringAllocator::prepareRead altogether just to fetch the hash.

Added benchmark for SetAccumulator to add 10M mostly unique values and read them
back. Without the optimizations, the benchmark couldn't finish within a
reasonable time (a few mininutes). Changing benchmark to process 100K values
allowed it to compelete.


Before (100K values):

```
============================================================================
[...]enchmarks/SetAccumulatorBenchmark.cpp     relative  time/iter   iters/s
============================================================================
bigint                                                      2.87ms    348.07
varchar                                                    22.22ms     45.01
twoBigints                                                988.08ms      1.01
```

After (100K values):

```
============================================================================
[...]enchmarks/SetAccumulatorBenchmark.cpp     relative  time/iter   iters/s
============================================================================
bigint                                                      2.80ms    356.87
varchar                                                    21.19ms     47.19
twoBigints                                                 38.83ms     25.76
```

After the optimizations, the original benchmark processing 10M values is 
finishing within a few seconds.


After (10M values):

```
============================================================================
[...]enchmarks/SetAccumulatorBenchmark.cpp     relative  time/iter   iters/s
============================================================================
bigint                                                       1.23s   814.20m
varchar                                                      2.96s   338.39m
twoBigints                                                   6.30s   158.70m
```